test(booking): move validBooking to utils/test-data.js
